[Do not commit, this is #626] Change how the various revisions states are handled in CompareWithBase
This is one of the most important commit in the set.

This removes the states displayedRevisionsXXXRevs. Instead the "in
progress" states are used to render the current view to the user.
The "staging" states are more like snapshots that the user can revert to
by pressing "cancel", or commit by pressing "save".