Bug 1949264 - Add git migration compatibility for ./mach try perf (#872)
* Bug 1949264 - Add git migration compatibility for ./mach try perf
As we are moving away from hg to git ./mach try perf we will no longer be able to get the try url and will be unable to generate perfcompare links.
This patch along with the 2 others on treeherder and mozilla central add a hash to the commit message and allows us to get a perfcompare url that finds the commits via that hash
This patch is still a work in progress as to properly test I will need the treeherder one to merge first to test properly as currently I am blocked with cross origin requests being not being allowed

* Bug 1949264 - Add git migration compatibility for ./mach try perf
As we are moving away from hg to git ./mach try perf we will no longer be able to get the try url and will be unable to generate perfcompare links.
This patch along with the 2 others on treeherder and mozilla central add a hash to the commit message and allows us to get a perfcompare url that finds the commits via that hash
This patch is still a work in progress as to properly test I will need the treeherder one to merge first to test properly as currently I am blocked with cross origin requests being not being allowed

* Bug 1949264 - Use UUID/markers to associate a base try run with a new try run

./mach try perf currently uses push-to-vcs to get a perma link to commits when they are pushed to try
When we migrate to github we will no longer have this and will not be able to create perma links for perfcompare
This perma link creation ability by ./mach try perf is a key reason for its use by many developers this issue must be fixed

This patch along with one on treeherder and one in mozilla central fixes the issue of no perma links with commit-based hashes
By creating a hash that won't collide we add a unique identifier to the commit messages when pushing with lando
This hash is then passed to perfcompare, where we can search for it by requesting it from the hash function from the tocommit api
When perfcompare gets the information on the revisions from the hashes it updates baseRev and newRev and the rest of perfcompare stays the same as if we were still using push-to-vcs

* Bug 1949264 - Use UUID/markers to associate a base try run with a new try run

./mach try perf currently uses push-to-vcs to get a perma link to commits when they are pushed to try
When we migrate to github we will no longer have this and will not be able to create perma links for perfcompare
This perma link creation ability by ./mach try perf is a key reason for its use by many developers this issue must be fixed

This patch along with one on treeherder and one in mozilla central fixes the issue of no perma links with commit-based hashes
By creating a hash that won't collide we add a unique identifier to the commit messages when pushing with lando
This hash is then passed to perfcompare, where we can search for it by requesting it from the hash function from the tocommit api
When perfcompare gets the information on the revisions from the hashes it updates baseRev and newRev and the rest of perfcompare stays the same as if we were still using push-to-vcs